// examples/remix/app/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Allocation = {
  __typename?: 'Allocation';
  /** If the Allocation is active it shows the indexer. If closed, it is null */
  activeForIndexer?: Maybe<Indexer>;
  /** Tokens allocation in this allocation */
  allocatedTokens: Scalars['BigInt']['output'];
  /** NOT IMPLEMENTED - Yearly annualzied return */
  annualizedReturn: Scalars['BigDecimal']['output'];
  /** Timestamp this allocation was closed at */
  closedAt?: Maybe<Scalars['Int']['output']>;
  /** Block hash at which this allocation was closed */
  closedAtBlockHash?: Maybe<Scalars['Bytes']['output']>;
  /** Block number at which this allocation was closed */
  closedAtBlockNumber?: Maybe<Scalars['Int']['output']>;
  /** Epoch this allocation was closed in */
  closedAtEpoch?: Maybe<Scalars['Int']['output']>;
  /** Timestamp this allocation was created at */
  createdAt: Scalars['Int']['output'];
  /** Block at which this allocation was created */
  createdAtBlockHash: Scalars['Bytes']['output'];
  /** Block number at which this allocation was created */
  createdAtBlockNumber: Scalars['Int']['output'];
  /** Epoch this allocation was created */
  createdAtEpoch: Scalars['Int']['output'];
  /** Creator of the allocation - can be the operator or the indexer */
  creator: Scalars['Bytes']['output'];
  /** Curator rewards deposited to the curating bonding curve */
  curatorRewards: Scalars['BigInt']['output'];
  /** Fees paid out to delegators */
  delegationFees: Scalars['BigInt']['output'];
  /** Query fee rebates collected from the protocol. Can differ from queryFeeRebates if multiple vouchers per allocation are allowed. */
  distributedRebates: Scalars['BigInt']['output'];
  /** [DEPRECATED] Effective allocation that is realized upon closing */
  effectiveAllocation: Scalars['BigInt']['output'];
  /** Channel Address */
  id: Scalars['ID']['output'];
  /** Indexer of this allocation */
  indexer: Indexer;
  /** Indexing rewards earned by this allocation by delegators */
  indexingDelegatorRewards: Scalars['BigInt']['output'];
  /** Indexing rewards earned by this allocation by indexers */
  indexingIndexerRewards: Scalars['BigInt']['output'];
  indexingRewardCutAtClose?: Maybe<Scalars['Int']['output']>;
  indexingRewardCutAtStart: Scalars['Int']['output'];
  indexingRewardEffectiveCutAtClose?: Maybe<Scalars['BigDecimal']['output']>;
  indexingRewardEffectiveCutAtStart: Scalars['BigDecimal']['output'];
  /** Indexing rewards earned by this allocation. Includes delegator and indexer rewards */
  indexingRewards: Scalars['BigInt']['output'];
  /** POI submitted with a closed allocation */
  poi?: Maybe<Scalars['Bytes']['output']>;
  /** [DEPRECATED] Pool in which this allocation was closed */
  poolClosedIn?: Maybe<Pool>;
  queryFeeCutAtClose?: Maybe<Scalars['Int']['output']>;
  queryFeeCutAtStart: Scalars['Int']['output'];
  queryFeeEffectiveCutAtClose?: Maybe<Scalars['BigDecimal']['output']>;
  queryFeeEffectiveCutAtStart: Scalars['BigDecimal']['output'];
  /** Query fee rebate amount claimed from the protocol through rebates mechanism. Does not include portion given to delegators */
  queryFeeRebates: Scalars['BigInt']['output'];
  /** Fees this allocation collected from query fees upon closing. Excludes curator reward and protocol tax */
  queryFeesCollected: Scalars['BigInt']['output'];
  /** Status of the allocation */
  status: AllocationStatus;
  /** Subgraph deployment that is being allocated to */
  subgraphDeployment: SubgraphDeployment;
  /** NOT IMPLEMENTED - Return for this allocation */
  totalReturn: Scalars['BigDecimal']['output'];
};
// ...
export type AuthorizedFunction = {
  __typename?: 'AuthorizedFunction';
  /** Function signature (string) */
  id: Scalars['ID']['output'];
  /** Token lock Manager */
  manager: TokenManager;
  /** Hash of the function signature */
  sigHash: Scalars['Bytes']['output'];
  /** The contract address that is authorized to have this function called on itself */
  target: Scalars['Bytes']['output'];
};
// ...
export type GraphAccount = {
  __typename?: 'GraphAccount';
  /** Graph token balance */
  balance: Scalars['BigInt']['output'];
  /** Balance received due to failed delegation transfer from L1 */
  balanceReceivedFromL1Delegation: Scalars['BigInt']['output'];
  /** Balance received due to failed signal transfer from L1 */
  balanceReceivedFromL1Signalling: Scalars['BigInt']['output'];
  bridgeDepositTransactions: Array<BridgeDepositTransaction>;
  bridgeWithdrawalTransactions: Array<BridgeWithdrawalTransaction>;
  /** Time the account was created */
  createdAt: Scalars['Int']['output'];
  /** Disputes this graph account has created */
  createdDisputes: Array<Dispute>;
  /** Amount this account has approved staking to transfer their GRT */
  curationApproval: Scalars['BigInt']['output'];
  /** Curator fields for this GraphAccount. Null if never curated */
  curator?: Maybe<Curator>;
  /** Default display name is the current default name. Used for filtered queries in the explorer */
  defaultDisplayName?: Maybe<Scalars['String']['output']>;
  /** Default name the graph account has chosen */
  defaultName?: Maybe<GraphAccountName>;
  /** Delegator fields for this GraphAccount. Null if never delegated */
  delegator?: Maybe<Delegator>;
  /** Time that this graph account became a developer */
  developerCreatedAt?: Maybe<Scalars['Int']['output']>;
  /** Disputes against this graph account */
  disputesAgainst: Array<Dispute>;
  /** Amount this account has approved the GNS to transfer their GRT */
  gnsApproval: Scalars['BigInt']['output'];
  /** Graph account ID */
  id: Scalars['ID']['output'];
  /** Indexer fields for this GraphAccount. Null if never indexed */
  indexer?: Maybe<Indexer>;
  metadata?: Maybe<GraphAccountMeta>;
  /** Name signal transactions created by this GraphAccount */
  nameSignalTransactions: Array<NameSignalTransaction>;
  /** All names this graph account has claimed from all name systems */
  names: Array<GraphAccountName>;
  /** Operator of other Graph Accounts */
  operatorOf: Array<GraphAccount>;
  /** Operators of this Graph Accounts */
  operators: Array<GraphAccount>;
  /** Amount this account has approved curation to transfer their GRT */
  stakingApproval: Scalars['BigInt']['output'];
  /** NOT IMPLEMENTED - Total query fees the subgraphs created by this account have accumulated in GRT */
  subgraphQueryFees: Scalars['BigInt']['output'];
  /** Subgraphs the graph account owns */
  subgraphs: Array<Subgraph>;
  tokenLockWallets: Array<TokenLockWallet>;
};
// ...
export type NameSignalSubgraphRelation = {
  __typename?: 'NameSignalSubgraphRelation';
  /** Subgraph ID + index */
  id: Scalars['ID']['output'];
  nameSignal: NameSignal;
  subgraph: Subgraph;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Access to subgraph metadata */
  _meta?: Maybe<_Meta_>;
  allocation?: Maybe<Allocation>;
  allocations: Array<Allocation>;
  attestation?: Maybe<Attestation>;
  attestations: Array<Attestation>;
  authorizedFunction?: Maybe<AuthorizedFunction>;
  authorizedFunctions: Array<AuthorizedFunction>;
  bridgeDepositTransaction?: Maybe<BridgeDepositTransaction>;
  bridgeDepositTransactions: Array<BridgeDepositTransaction>;
  bridgeWithdrawalTransaction?: Maybe<BridgeWithdrawalTransaction>;
  bridgeWithdrawalTransactions: Array<BridgeWithdrawalTransaction>;
  curator?: Maybe<Curator>;
  curatorSearch: Array<Curator>;
  curators: Array<Curator>;
  currentSubgraphDeploymentRelation?: Maybe<CurrentSubgraphDeploymentRelation>;
  currentSubgraphDeploymentRelations: Array<CurrentSubgraphDeploymentRelation>;
  delegatedStake?: Maybe<DelegatedStake>;
  delegatedStakes: Array<DelegatedStake>;
  delegator?: Maybe<Delegator>;
  delegatorSearch: Array<Delegator>;
  delegators: Array<Delegator>;
  dispute?: Maybe<Dispute>;
  disputes: Array<Dispute>;
  epoch?: Maybe<Epoch>;
  epoches: Array<Epoch>;
  graphAccount?: Maybe<GraphAccount>;
  graphAccountMeta?: Maybe<GraphAccountMeta>;
  graphAccountMetas: Array<GraphAccountMeta>;
  graphAccountName?: Maybe<GraphAccountName>;
  graphAccountNames: Array<GraphAccountName>;
  graphAccounts: Array<GraphAccount>;
  graphNetwork?: Maybe<GraphNetwork>;
  graphNetworks: Array<GraphNetwork>;
  indexer?: Maybe<Indexer>;
  indexers: Array<Indexer>;
  nameSignal?: Maybe<NameSignal>;
  nameSignalSubgraphRelation?: Maybe<NameSignalSubgraphRelation>;
  nameSignalSubgraphRelations: Array<NameSignalSubgraphRelation>;
  nameSignalTransaction?: Maybe<NameSignalTransaction>;
  nameSignalTransactions: Array<NameSignalTransaction>;
  nameSignals: Array<NameSignal>;
  pool?: Maybe<Pool>;
  pools: Array<Pool>;
  retryableTicket?: Maybe<RetryableTicket>;
  retryableTicketRedeemAttempt?: Maybe<RetryableTicketRedeemAttempt>;
  retryableTicketRedeemAttempts: Array<RetryableTicketRedeemAttempt>;
  retryableTickets: Array<RetryableTicket>;
  signal?: Maybe<Signal>;
  signalTransaction?: Maybe<SignalTransaction>;
  signalTransactions: Array<SignalTransaction>;
  signals: Array<Signal>;
  subgraph?: Maybe<Subgraph>;
  subgraphDeployment?: Maybe<SubgraphDeployment>;
  subgraphDeploymentManifest?: Maybe<SubgraphDeploymentManifest>;
  subgraphDeploymentManifests: Array<SubgraphDeploymentManifest>;
  subgraphDeploymentSchema?: Maybe<SubgraphDeploymentSchema>;
  subgraphDeploymentSchemas: Array<SubgraphDeploymentSchema>;
  subgraphDeployments: Array<SubgraphDeployment>;
  subgraphMeta?: Maybe<SubgraphMeta>;
  subgraphMetadataSearch: Array<SubgraphMeta>;
  subgraphMetas: Array<SubgraphMeta>;
  subgraphVersion?: Maybe<SubgraphVersion>;
  subgraphVersionMeta?: Maybe<SubgraphVersionMeta>;
  subgraphVersionMetas: Array<SubgraphVersionMeta>;
  subgraphVersions: Array<SubgraphVersion>;
  subgraphs: Array<Subgraph>;
  tokenLockWallet?: Maybe<TokenLockWallet>;
  tokenLockWallets: Array<TokenLockWallet>;
  tokenManager?: Maybe<TokenManager>;
  tokenManagers: Array<TokenManager>;
  transaction?: Maybe<Transaction>;
  transactions: Array<Transaction>;
};
// ...
export type SubgraphMeta = {
  __typename?: 'SubgraphMeta';
  /** Categories that the subgraph belongs to. */
  categories?: Maybe<Array<Scalars['String']['output']>>;
  /** Location of the code for this project */
  codeRepository?: Maybe<Scalars['String']['output']>;
  /** Short description of the subgraph */
  description?: Maybe<Scalars['String']['output']>;
  /** Display name */
  displayName?: Maybe<Scalars['String']['output']>;
  /** Subgraph metadata ipfs hash */
  id: Scalars['ID']['output'];
  /** Image in string format */
  image?: Maybe<Scalars['String']['output']>;
  /** NFT Image representation */
  nftImage?: Maybe<Scalars['String']['output']>;
  /** Subgraph that reference this metadata. For compatibility purposes. For the full list use subgraphs */
  subgraph?: Maybe<Subgraph>;
  /** Subgraphs that reference this metadata */
  subgraphs: Array<Subgraph>;
  /** Projects website */
  website?: Maybe<Scalars['String']['output']>;
};
// ...
export type _Meta_ = {
  __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   *
   */
  block: _Block_;
  /** The deployment ID */
  deployment: Scalars['String']['output'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  hasIndexingErrors: Scalars['Boolean']['output'];
};

export type _SubgraphErrorPolicy_ =
  /** Data will be returned even if the subgraph has indexing errors */
  | 'allow'
  /** If the subgraph has indexing errors, data will be omitted. The default. */
  | 'deny';

export type SubgraphsQueryVariables = Exact<{
  first?: InputMaybe<Scalars['Int']['input']>;
  skip?: InputMaybe<Scalars['Int']['input']>;
  where?: InputMaybe<Subgraph_Filter>;
}>;

export type SubgraphsQuery = {
  __typename?: 'Query';
  subgraphs: Array<{
    __typename?: 'Subgraph';
    id: string;
    metadata?: {
      __typename?: 'SubgraphMeta';
      displayName?: string | null;
      image?: string | null;
      description?: string | null;
    } | null;
    owner: { __typename?: 'GraphAccount'; id: string };
  }>;
};